TP-198 wip(CrudTestesUnitarios): feitos ajustes nos tipos de dado para melhor adaptação ao contexto. Onde haviam valores Double, foram todos alterados para BigDecimal. Do mesmo modo, houve ajuste no MSSQL para receber dados Decimal(10, 2)

+++ crud/src/main/java/com/agrotis/trainees/crud/dto/NotaFiscalDto.java
@@ -1,5 +1,6 @@
 package com.agrotis.trainees.crud.dto;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.util.List;
 import java.util.stream.Collectors;
@@ -17,7 +18,7 @@ public class NotaFiscalDto {
     private Integer numero;
     private LocalDate data;
     private List<NotaFiscalItem> itens;
-    private Double valorTotal;
+    private BigDecimal valorTotal;
 
     public NotaFiscalDto() {
         super();
@@ -74,11 +75,11 @@ public class NotaFiscalDto {
         this.itens = itens;
     }
 
-    public Double getValorTotal() {
+    public BigDecimal getValorTotal() {
         return valorTotal;
     }
 
-    public void setValorTotal(Double valorTotal) {
+    public void setValorTotal(BigDecimal valorTotal) {
         this.valorTotal = valorTotal;
     }
